refactored ViewAllBody into ViewAllPage
